requirements.txt
discord.py
python-dotenv
yt-dlp
legacy-cgi
PyNaCl